move post-process into src/fuzz_introspector
