+++ src/fuzz_introspector/analyses/bug_digestor.py
@@ -20,12 +20,12 @@ from typing import (
     Tuple,
 )
 
-import analysis
-import data_loader
-import html_helpers
+from fuzz_introspector import analysis
+from fuzz_introspector import data_loader
+from fuzz_introspector import html_helpers
 
-import datatypes.project_profile
-import datatypes.fuzzer_profile
+from fuzz_introspector import datatypes.project_profile
+from fuzz_introspector import datatypes.fuzzer_profile
 
 logger = logging.getLogger(name=__name__)
