+++ src/fuzz_introspector/analyses/calltree_analysis.py
@@ -24,13 +24,13 @@ from typing import (
     Set,
 )
 
-import analysis
-import utils
-import cfg_load
-import html_helpers
+from fuzz_introspector import analysis
+from fuzz_introspector import utils
+from fuzz_introspector import cfg_load
+from fuzz_introspector import html_helpers
 
-import datatypes.project_profile
-import datatypes.fuzzer_profile
+from fuzz_introspector import datatypes.project_profile
+from fuzz_introspector import datatypes.fuzzer_profile
 
 # For pretty printing the html code:
 from bs4 import BeautifulSoup as bs
