+++ src/fuzz_introspector/analyses/filepath_analyser.py
@@ -22,11 +22,10 @@ from typing import (
     Tuple
 )
 
-import analysis
-import html_helpers
-
-import datatypes.project_profile
-import datatypes.fuzzer_profile
+from fuzz_introspector import analysis
+from fuzz_introspector import html_helpers
+from fuzz_introspector import datatypes.project_profile
+from fuzz_introspector import datatypes.fuzzer_profile
 
 logger = logging.getLogger(name=__name__)
