+++ src/fuzz_introspector/constants.py
