+++ src/fuzz_introspector/datatypes/bug.py
