+++ src/fuzz_introspector/html_report.py
@@ -27,14 +27,14 @@ from typing import (
     Tuple,
 )
 
-import analysis
-import utils
-import cfg_load
-import constants
-import html_helpers
+from fuzz_introspector import analysis
+from fuzz_introspector import utils
+from fuzz_introspector import cfg_load
+from fuzz_introspector import constants
+from fuzz_introspector import html_helpers
 
-import datatypes.project_profile
-import datatypes.fuzzer_profile
+from fuzz_introspector import datatypes.project_profile
+from fuzz_introspector import datatypes.fuzzer_profile
 
 # For pretty printing the html code:
 from bs4 import BeautifulSoup as bs
