+++ src/fuzz_introspector/styling/calltree.js
