+++ src/fuzz_introspector/styling/clike.js
